+++ source-code/clj/x/server_developer/api.clj
@@ -1,21 +1,17 @@
 
-;; -- Header ------------------------------------------------------------------
-;; ----------------------------------------------------------------------------
-
-; Author: bithandshake
-; Created: 2021.01.02
-; Description:
-; Version: v0.2.8
-; Compatibility: x4.5.5
-
-
-
 ;; -- Namespace ---------------------------------------------------------------
 ;; ----------------------------------------------------------------------------
 
 (ns x.server-developer.api
-    (:require [x.server-developer.developer-tools]
-              [x.server-developer.mongo-browser]
-              [x.server-developer.re-frame-browser]
-              [x.server-developer.docs]
-              [x.server-developer.engine :as engine]))
+    (:require [x.server-developer.developer-tools.lifecycles]
+              [x.server-developer.developer-tools.routes]
+              [x.server-developer.developer-tools.styles]
+              [x.server-developer.developer-tools.views]
+              [x.server-developer.mongo-db-browser.styles]
+              [x.server-developer.mongo-db-browser.views]
+              [x.server-developer.re-frame-browser.styles]
+              [x.server-developer.re-frame-browser.views]
+              [x.server-developer.docs.engine]
+              [x.server-developer.docs.lifecycles]
+              [x.server-developer.docs.routes]
+              [x.server-developer.side-effects]))