+++ source-code/clj/x/server_developer/developer_tools/styles.clj
@@ -1,0 +1,19 @@
+
+;; -- Namespace ---------------------------------------------------------------
+;; ----------------------------------------------------------------------------
+
+(ns x.server-developer.developer-tools.styles)
+
+
+
+;; -- Styles ------------------------------------------------------------------
+;; ----------------------------------------------------------------------------
+
+(def body-style "padding: 0; margin: 0")
+
+(def menu-bar-style (str "display: flex; width: 100%; background-color: #d5c1ef;"))
+
+(def menu-item-style (str "text-decoration: none; padding: 0 16px; line-height: 32px; display: block;"
+                          "background-color: #cbf1dc; margin: 8px; border-radius: 16px; "
+                          "font-weight: 500; font-size: 13px; color: #222 !important;"
+                          " border: 1px solid #888"))