+++ source-code/clj/x/server_developer/mongo_db_browser/styles.clj
@@ -1,0 +1,20 @@
+
+;; -- Namespace ---------------------------------------------------------------
+;; ----------------------------------------------------------------------------
+
+(ns x.server-developer.mongo-db-browser.styles)
+
+
+
+;; -- Styles ------------------------------------------------------------------
+;; ----------------------------------------------------------------------------
+
+(def button-wrapper-style (str "display: flex; background-color: #d5c1ef;"))
+
+(def button-style (str "text-decoration: none; padding: 0 16px; line-height: 32px; display: flex;"
+                       "border-radius: 16px; margin: 8px; background: #cbf1dc; color: #222 !important; "
+                       "font-size: 13px; font-weight: 500; border: 1px solid #888"))
+
+(def danger-button-style (str "text-decoration: none; padding: 0 16px; line-height: 32px; display: flex;"
+                              "border-radius: 16px; margin: 8px; background: #f3a9e3; color: #222 !important; "
+                              "font-size: 13px; font-weight: 500; border: 1px solid #888"))